### 2018-10-12 #982 @arfon
@trallard - would you be able to edit this submission for JOSS?
