### 2018-10-12 #982 @trallard
Sure @arfon

👋 @SimonGreenhill I will be handling editor for your submission. Have you thought about any reviewers for the package?
If you have not gotten anyone in mind you can check the  current reviewers registered using the following command `@whedon list reviewers`
