### 2018-10-12 #982 @SimonGreenhill
Thanks @trallard, any of the python people should be ok. It's a very simple package* so should be quite easy to review, I hope

(* in fact, I worry it's _too_ simple, but it does fill a small and annoying gap in the toolkit).
