### 2018-10-12 #982 @arfon
> @whedon assign @trallard as editor

Looks like @whedon is sleepy... Let me check in on this...
