### 2018-10-15 #982 @trallard
Hi @RichardLitt @deniederhut @krother  @genomematt would any of you be interested in and have availability to review this JOSS submission?
We'd ideally be looking at getting two reviewers
