### 2018-10-23 #982 @trallard
Closing this issue now. @deniederhut  and @RichardLitt let's proceed with the review at #1040  🌟
