### 2018-10-31 SimonGreenhill/treemaker#2 @deniederhut
Paper: references missing DOIs
There are two papers in the references section (the one on the Nexus format, and the one on the Papuan New Guinea languages) that are missing DOIs. It looks like Ethnologue does not have a persistent identifier associated with its software product, so it might be nice to have a URL for it.

xref https://github.com/openjournals/joss-reviews/issues/1040
