### 2018-10-31 f4e320e
Adding DOIs and URLs
