+++ paper/paper.bib
@@ -5,7 +5,9 @@
  pages = {590-621},
  title = {Nexus: {{An}} Extensible File Format for Systematic Information},
  volume = {46},
- year = {1997}
+ year = {1997},
+ doi = {10.1093/sysbio/46.4.590},
+ url = {https://doi.org/10.1093/sysbio/46.4.590}
 }
 
 @book{Ethnologue,
@@ -14,7 +16,8 @@
  editor = {Simons, Gary F. and Charles D. Fennig},
  publisher = {{SIL International}},
  title = {Ethnologue: {{Languages}} of the {{World}}},
- year = {2009}
+ year = {2009},
+ url = {http://ethnologue.com/}
 }
 
 @incollection{Wurm1975,
@@ -24,7 +27,9 @@
   editor = {Wurm, S. A.},
   publisher = {{Pacific Linguistics}},
   title = {The {{East Papuan Phylum}} in {{General}}},
-  year = {1975}
+  year = {1975},
+  url = {https://openresearch-repository.anu.edu.au/bitstream/1885/145150/1/PL-C38.pdf},
+  doi = {http://dx.doi.org/10.15144/PL-C38}
 }
 
 @article{Bromham2018,
