### 2018-10-31 SimonGreenhill/treemaker#3 @deniederhut
DOC: request for additional documentation
In particular, I am not easily able to find:

1. installation instructions
2. API documentation
3. Contributing guide

#### 1. Install

At a minimum, I would like to see `git clone <> && python setup.py install`. It would be nice to have instructions for a virtual environment manager (`pipenv`, `conda`, or `virtualenv`) and a package manager (`pip`, or `poetry`). It would be great if this was submitted to pypi (see [this guide](https://packaging.python.org/tutorials/packaging-projects/#uploading-the-distribution-archives)).

#### 2. API documentation

It looks like most of your objects have docstrings attached, so `sphinx` should be able to build this for you.

#### 3. Contributing guide

GitHub will automatically warn new contributors to look at the guide before submitting their first issue/PR -- you can see how to set this up [here](https://help.github.com/articles/setting-guidelines-for-repository-contributors/). At a minimum, I would like to see short descriptions of how you would like people to report issues / submit code for inclusion. It can be nice (for you, if this package should become wildly popular) to include issue templates. In a perfect world there would also be a code of conduct.

Please feel free to reach out to me directly if you would like guidance on any of this.

xref https://github.com/openjournals/joss-reviews/issues/1040
