### 2018-10-31 b54457b
Create CONTRIBUTING.md
