### 2018-10-31 5eaf75d
Update CONTRIBUTING.md
