### 2018-10-31 221293e
Update CONTRIBUTING.md
