+++ CONTRIBUTING.md
@@ -60,7 +60,7 @@ further defined and clarified by project maintainers.
 ## Enforcement
 
 Instances of abusive, harassing, or otherwise unacceptable behavior may be
-reported by contacting the project team at <simon at simon.net.nz>. All
+reported by contacting the project team at the email address [listed here](https://github.com/SimonGreenhill/treemaker/blob/master/setup.py#L27). All
 complaints will be reviewed and investigated and will result in a response that
 is deemed necessary and appropriate to the circumstances. The project team is
 obligated to maintain confidentiality with regard to the reporter of an incident.
