### 2018-10-31 SimonGreenhill/treemaker#3 @SimonGreenhill
Added Contributing Guide in https://github.com/SimonGreenhill/treemaker/commit/221293e283d1fd72ebdc698b8c58e97116b182db
