### 2018-10-31 SimonGreenhill/treemaker#3 @SimonGreenhill
It's on PyPI [here](https://pypi.org/project/treemaker/), and I've documented install procedure in https://github.com/SimonGreenhill/treemaker/commit/e2df4694b3972241d7a13d6e730dc1cfc71c4dae
