+++ treemaker/treemaker.py
@@ -2,7 +2,7 @@
 #coding=utf-8
 """TreeMaker"""
 __author__ = 'Simon J. Greenhill <simon@simon.net.nz>'
-__copyright__ = 'Copyright (c) 2016 Simon J. Greenhill'
+__copyright__ = 'Copyright (c) 2018 Simon J. Greenhill'
 __license__ = 'New-style BSD'
 
 import os
@@ -10,8 +10,9 @@ import re
 import sys
 import codecs
 import argparse
+from functools import total_ordering
 
-VERSION = "1.0.3"
+VERSION = "1.1"
 
 NEXUS_TEMPLATE = """#NEXUS
 
@@ -23,7 +24,15 @@ end;
 IS_WHITESPACE = re.compile(r"""\s+""")
 
 
+@total_ordering
 class Tree(object):
+    """
+    Tree object to represent the classification taxonomy.
+
+    Args:
+        node (str): Label for this node.
+        children (list): Optional list of children nodes
+    """
     def __init__(self, node=None, children=None):
         self.children = []
         
@@ -37,9 +46,6 @@ class Tree(object):
     
     def __lt__(self, other):
         return self.node < other.node
-        
-    def __gt__(self, other):
-        return self.node > other.node
     
     def __eq__(self, other):
         return self.node == other.node
@@ -55,21 +61,48 @@ class Tree(object):
         return len(self.children) > 0
     
     def add(self, node, children=None):
-        """Adds a node (with optional children) to the tree"""
+        """
+        Adds a node (with optional children) to the tree.
+
+        Args:
+            node (str): Label for this node.
+            children (list): Optional list of children nodes
+
+        Returns:
+            treemaker.Tree: the created node matching `label`.
+        """
         if not isinstance(node, Tree):
             node = Tree(node, children)
         self.children.append(node)
         return node
     
     def get_or_create(self, label):
-        """Helper function to get or create a node"""
+        """
+        Helper function to get or create a node.
+
+        Args:
+            label (str): Label for this node.
+
+        Returns:
+            treemaker.Tree: the found or created node matching `label`.
+        """
         found = self.get(label)
         if found:
             return found
         return self.add(label)
     
     def get(self, label, node=None):
-        """Searches tree for node `label`"""
+        """
+        Searches tree for node `label`.
+
+        Args:
+            label (str): Label for this node.
+            node (treemaker.Tree): (optional) parent node.
+                Default is current node.
+
+        Returns:
+            treemaker.Tree: the found or created node matching `label`.
+        """
         node = self if node is None else node
         for child in node.children:
             if child.node == label:
@@ -80,7 +113,16 @@ class Tree(object):
         return None
     
     def tips(self, node=None):
-        """Returns a list of the tips in the tree"""
+        """
+        Returns a list of the tips in the tree
+
+        Args:
+            node (treemaker.Tree): (optional) parent node. Default is current 
+                node.
+
+        Returns:
+            List[treemaker.Tree]: the tip nodes from the given node.
+        """
         node = self if node is None else node
         for child in node.children:
             if child.is_node:  # remove yield from so we can support py2.7
@@ -111,11 +153,24 @@ class TreeMaker(object):
     def _check_taxon(self, taxon):
         for char in "()":
             if char in taxon:
-                raise ValueError("Error: %s is not allowed in taxon names" % char)
+                raise ValueError(
+                    "Error: %s is not allowed in taxon names" % char
+                )
     
     def add(self, leaf, classification):
         """
         Adds `leaf` to the tree in the location specified by `classification`
+
+        Args:
+            leaf (str): Leaf label
+            classification (str): A classification string of a format handled
+                by `parse_classification`.
+
+        Returns:
[…]
