+++ docs/build/html/_sources/index.rst.txt
@@ -1,0 +1,26 @@
+Welcome to treemaker's documentation!
+=====================================
+
+.. toctree::
+   :maxdepth: 3
+   :caption: Contents:
+
+.. autoclass:: treemaker.Tree
+   :members:
+   :undoc-members:
+   :inherited-members:
+   :show-inheritance:
+
+.. autoclass:: treemaker.TreeMaker
+   :members:
+   :undoc-members:
+   :inherited-members:
+   :show-inheritance:
+
+
+Indices and tables
+==================
+
+* :ref:`genindex`
+* :ref:`modindex`
+* :ref:`search`
