+++ docs/build/html/_sources/modules.rst.txt
@@ -1,0 +1,7 @@
+treemaker
+=========
+
+.. toctree::
+   :maxdepth: 4
+
+   treemaker
