+++ docs/source/treemaker.rst
@@ -1,0 +1,30 @@
+treemaker package
+=================
+
+Submodules
+----------
+
+treemaker.test\_treemaker module
+--------------------------------
+
+.. automodule:: treemaker.test_treemaker
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+treemaker.treemaker module
+--------------------------
+
+.. automodule:: treemaker.treemaker
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+
+Module contents
+---------------
+
+.. automodule:: treemaker
+    :members:
+    :undoc-members:
+    :show-inheritance:
