+++ docs/requirements.txt
@@ -1,0 +1,3 @@
+sphinx
+sphinxcontrib-napoleon
+
