+++ MANIFEST.in
@@ -1,5 +1,6 @@
-include DESCRIPTION.rst
 include LICENSE
+include CONTRIBUTING.md
+include README.md
 exclude .gitignore
 exclude .coverage
 exclude .travis.yml
@@ -7,5 +8,4 @@ prune .cache
 prune .git
 prune build
 prune dist
-recursive-include tests *
-
+prune paper
