### 2018-11-01 5489dad
Pin sphinx to latest version 1.8.1
