+++ docs/requirements.txt
@@ -1,3 +1,3 @@
-sphinx
+sphinx==1.8.1
 sphinxcontrib-napoleon
 
