### 2018-11-01 8a7bdea
Pin sphinxcontrib-napoleon to latest version 0.7
