+++ docs/requirements.txt
@@ -1,3 +1,3 @@
 sphinx
-sphinxcontrib-napoleon
+sphinxcontrib-napoleon==0.7
 
