### 2018-11-01 SimonGreenhill/treemaker#3 @SimonGreenhill
And here's the [API documentation](http://simongreenhill.github.io/treemaker/), committed in https://github.com/SimonGreenhill/treemaker/commit/226071c7b0a4879095bd744f8ea1ed9b14a8e8f1
