### 2018-11-01 #1040 @deniederhut
@trallard the most excellent @SimonGreenhill has addressed the issues I've filed, and I believe this library is ready for inclusion in JOSS
