### 2018-11-01 SimonGreenhill/treemaker#12 @RichardLitt
Increase test coverage to 100%
You're one test away from 100% code coverage. I think it might be worth adding another test. :)

https://coveralls.io/builds/19852503/source?filename=treemaker/treemaker.py#L210

Crossref openjournals/joss-reviews#1040
