### 2018-11-01 ace4ed0
removing some excess code
