### 2018-11-01 SimonGreenhill/treemaker#12 @SimonGreenhill
Done in https://github.com/SimonGreenhill/treemaker/commit/78c354c97dba64afde893b976ede76d0d08851b1 and https://github.com/SimonGreenhill/treemaker/commit/ace4ed06c7a86c85593f728e98242372d088ea8d
