### 2018-11-01 214dcef
Adding JOSS badge
