### 2018-11-01 #1040 @trallard
Thanks @deniederhut for your time as a reviewer!

I see there are still some unchecked items from @RichardLitt 's list. I will wait until I hear from him to proceed to do some editorial checks on the submission 😄
