### 2018-11-01 c3a702f
adding details from paper.md to readme
