+++ README.md
@@ -7,6 +7,63 @@ A Python library for creating a Newick formatted tree from a set of classificati
 [![DOI](https://zenodo.org/badge/22704/SimonGreenhill/treemaker.svg)](https://zenodo.org/badge/latestdoi/22704/SimonGreenhill/treemaker)
 [![status](http://joss.theoj.org/papers/19eae6958062fc8a72d8a02efdaf8b23/status.svg)](http://joss.theoj.org/papers/19eae6958062fc8a72d8a02efdaf8b23)
 
+
+
+Often research in linguistics or cultural evolution produces tree taxonomies or
+classifications. However, these are often not in a format readily available for
+use in programs that can understand and manipulate trees. For example, the
+global taxonomy of languages published by the
+[Ethnologue](https://www.ethnologue.com/) [@Ethnologue] classifies languages
+into families and subgroups using a taxonomy string e.g. the language
+[Kalam](https://www.ethnologue.com/language/kmh) is classified as "Trans-New
+Guinea, Madang, Kalam-Kobon", while
+[Mauwake](https://www.ethnologue.com/language/mhl) is classified as "Trans-New
+Guinea, Madang, Croisilles, Pihom", and
+[Kare](https://www.ethnologue.com/language/kmf) is "Trans-New Guinea, Madang,
+Croisilles, Kare". This classification indicates that while all these languages
+are part of the Madang subgroup of the Trans-New Guinea language family, Kare
+and Mauwake are more closely related (as they belong to the Croisilles
+subgroup).
+
+Other publications use a tabular indented format to demarcate relationships,
+such as the example in Figure 1 from Stephen Wurm's classification of his
+proposed Yele-Solomons language phylum (Wurm 1975).
+
+Both the taxonomy string and tabular format however are hard to load into
+software packages that can analyse, compare, visualise and manipulate trees.
+```treemaker``` aims to make this easy by converting taxonomic data into
+[Newick](https://en.wikipedia.org/wiki/Newick_format) and Nexus (Maddison 1997)
+formats commonly used by phylogenetic manipulation programs.
+
+## Converting a Taxonomy to a Tree:
+
+```treemaker``` can convert a text file with a taxonomy (easily obtained from
+Ethnologue or manually entered) like this:
+
+```text
+Bilua       Yele-Solomons, Central Solomon
+Baniata     Yele-Solomons, Central Solomon
+Lavukaleve  Yele-Solomons, Central Solomon
+Savosavo    Yele-Solomons, Central Solomon
+Kazukuru    Yele-Solomons, Kazukuru
+Guliguli    Yele-Solomons, Kazukuru
+Dororo      Yele-Solomons, Kazukuru
+Yele        Yele-Solomons
+```
+
+Into a Newick tree representation:
+
+```text
+((Baniata,Bilua,Lavukaleve,Savosavo),(Dororo,Guliguli,Kazukuru),Yele);
+```
+
+...which can then be loaded into phylogenetic programs to visualise or manipulate.
+
+```treemaker``` has been used to enable the analyses in Bromham et al. (2018), and a number of 
+forthcoming articles.
+
+
+
 ## Installation:
 
 Installation is only a pip install away:
@@ -124,6 +181,12 @@ To run treemaker's tests simply run:
 > python treemaker/test_treemaker.py
 ```
 
-## Support
+## Support:
 
 For questions on how to use or update this, feel free to [open an issue](https://github.com/SimonGreenhill/treemaker/issues). I'll get to it as soon as I can. 
+
+## References:
+
[…]
