### 2018-11-01 #1040 @RichardLitt
@trallard Well, @SimonGreenhill just finished doing a fantastically fast and thorough job resolving my niggling complaints. I think this is good to publish as it is now. Thanks, Simon! Cool package.
