### 2018-11-04 #1040 @trallard
@SimonGreenhill  I just checked the [DOI](https://zenodo.org/record/55660#.W97vMmj7RPY) you created for your repository and this resolves correctly

I will regenerate the pdf to do some basic checks as this seems to not have worked properly (?)
