### 2018-11-07 #1040 @SimonGreenhill
Thanks for the enthusiasm, whedon. I've checked the proof and made a few minor tweaks.
