### 2018-11-08 #1040 @trallard
@SimonGreenhill gentle ping. I think this is there for acceptance but I would like to check about the DOI version. Would you be interested in uploading/sync to `1.0.3` so that the version here and the DIO versions match?
I also noticed that in Zendo the package name is _Treemaker v1_ you might want to edit it to remove the v1
