### 2018-11-08 5d26704
updating version to 1.2 due to mismatches in different places
