### 2018-11-08 #1040 @SimonGreenhill
Ok, I realised there were mismatching version numbers (1.1 and 1.03) so I've released a new version 1.2, and the [Zenodo version](https://zenodo.org/record/1480504) has been updated.

I think this fixes the "Treemaker v1" issue too -- I can't see it. Could you please double-check for me?
