### 2018-11-08 #1040 @trallard
@arfon : This submission is accepted and ready to be published 🎉

@RichardLitt @deniederhut thank you very much for your time and valuable contribution to JOSS as reviewers for this submission
