### 2018-11-08 #1040 @RichardLitt
Thanks for the opportunity to learn more about trees. 🌲
