### 2018-11-08 #1040 @arfon
@SimonGreenhill - your paper is now accepted into JOSS ⚡️🚀💥
