### 2018-11-08 #1040 @SimonGreenhill
Awesome, thanks everyone, the review process really made this package better!
